opt-cold-rolling/pyproject.toml
[project]
name = "opt-cold-rolling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deap>=1.4.3",
    "dotenv>=0.9.9",
    "ipykernel>=7.0.1",
    "jupyter>=1.1.1",
    "kaleido>=1.1.0",
    "notebook>=7.4.7",
    "numpy>=2.3.4",
    "pandas-stubs==2.3.2.250926",
]
